requirements.txt
python-docx==1.1.0
openai>=1.40.0
python-dotenv==1.0.0
streamlit>=1.31.0
numpy>=1.26.0

